fix(azure-speech-service): update private endpoint location handling (#88)
This commit modifies the Azure Speech Service module to use the
`vnet_location` parameter for private endpoints instead of the cognitive
account's location. Additionally, the module version is bumped to 3.0.0,
and compatibility requirements are updated in the module.yaml file. The
README and variable definitions are also updated to reflect these
changes.


## Checklist before requesting a review
- [x] Module version bumped (`module.yaml`)
- [x] Changelog updated and where applicable compatibility changed
(`module.yaml`)
- [x] Compatibility updated where applicable  (`README.md`)
- [x] Pre-Commit passed or has been run manually (`Makefile`)

---------

Co-authored-by: Dominik from Unique <dominik@unique.ch>