feat(azure-storage-account): add enable_backup_role_assignment variab… (#94)
add enable_backup_role_assignment variable and update role assignment
logic

This commit introduces a new variable `enable_backup_role_assignment` to
control the creation of the role assignment for the backup vault
accessing the storage account. The role assignment logic is updated to
check this variable, enhancing flexibility in managing permissions.
Additionally, a new output for the backup vault's managed identity
principal ID is added to the outputs file, and the README is updated to
reflect these changes.

## Describe your changes

<!-- Describe your changes in detail -->

## Checklist before requesting a review
- [x] Module version bumped (`module.yaml`)
- [x] Changelog updated and where applicable compatibility changed
(`module.yaml`)
- [x] Compatibility updated where applicable  (`README.md`)
- [x] Pre-Commit passed or has been run manually (`Makefile`)